refactor(campaignPage): hoist shared chrome to shared.jsx — renderer cycle + TDZ workaround gone (P4-7/1)
